# docs/openapi.swagger.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "AI Audit System API",
                 "version":  "2.3.0",
                 "description":  "REST API for AI-powered quality audit system with GPT-5",
                 "contact":  {
                                 "name":  "API Support"
                             }
             },
    "servers":  [
                    {
                        "url":  "http://localhost:3002",
                        "description":  "Development server"
                    }
                ],
...
    "paths":  {
...
                  "/api/webhooks/test":  {
                                             "post":  {
                                                          "summary":  "Test webhook delivery",
                                                          "description":  "Send test webhook events to the frontend for testing purposes",
                                                          "tags":  [
                                                                       "Webhooks (Testing)"
                                                                   ],
                                                          "requestBody":  {
                                                                              "required":  true,
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "object",
                                                                                                                                      "properties":  {
                                                                                                                                                         "eventType":  {
                                                                                                                                                                           "type":  "string",
                                                                                                                                                                           "enum":  [
                                                                                                                                                                                        "audit.started",
                                                                                                                                                                                        "audit.progress",
                                                                                                                                                                                        "audit.completed",
                                                                                                                                                                                        "audit.failed",
                                                                                                                                                                                        "transcription.progress",
                                                                                                                                                                                        "transcription.completed",
                                                                                                                                                                                        "batch.progress",
                                                                                                                                                                                        "batch.completed",
                                                                                                                                                                                        "notification",
                                                                                                                                                                                        "all"
                                                                                                                                                                                    ],
                                                                                                                                                                           "description":  "Type of webhook event to send (or \u0027all\u0027 to send all types)"
                                                                                                                                                                       }
                                                                                                                                                     }
                                                                                                                                  },
                                                                                                                       "examples":  {
                                                                                                                                        "single":  {
                                                                                                                                                       "value":  {
                                                                                                                                                                     "eventType":  "notification"
                                                                                                                                                                 }
                                                                                                                                                   },
                                                                                                                                        "all":  {
                                                                                                                                                    "value":  {
                                                                                                                                                                  "eventType":  "all"
                                                                                                                                                              }
                                                                                                                                                }
                                                                                                                                    }
                                                                                                                   }
                                                                                          }
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "Webhook(s) sent successfully"
                                                                                    },
                                                                            "400":  {
                                                                                        "description":  "Invalid event type"
                                                                                    },
                                                                            "500":  {
                                                                                        "description":  "Webhook delivery failed"
                                                                                    }
                                                                        }
                                                      }
                                         },
                  "/api/webhooks/test/workflow":  {
                                                      "post":  {
                                                                   "summary":  "Simulate a complete audit workflow",
                                                                   "description":  "Sends a sequence of webhooks simulating a full audit lifecycle",
                                                                   "tags":  [
                                                                                "Webhooks (Testing)"
                                                                            ],
                                                                   "requestBody":  {
                                                                                       "required":  false,
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "type":  "object",
                                                                                                                                               "properties":  {
                                                                                                                                                                  "delay":  {
                                                                                                                                                                                "type":  "number",
                                                                                                                                                                                "description":  "Delay in milliseconds between events (default 1000)",
                                                                                                                                                                                "example":  2000
                                                                                                                                                                            }
                                                                                                                                                              }
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Workflow simulation completed"
                                                                                             }
                                                                                 }
                                                               }
                                                  },
                  "/api/webhooks/test/custom":  {
                                                    "post":  {
                                                                 "summary":  "Send custom webhook event",
                                                                 "description":  "Send a custom webhook with any event type and data",
                                                                 "tags":  [
                                                                              "Webhooks (Testing)"
                                                                          ],
                                                                 "requestBody":  {
                                                                                     "required":  true,
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "type":  "object",
                                                                                                                                             "properties":  {
                                                                                                                                                                "event":  {
                                                                                                                                                                              "type":  "string",
                                                                                                                                                                              "description":  "Event type",
                                                                                                                                                                              "example":  "audit.progress"
                                                                                                                                                                          },
                                                                                                                                                                "data":  {
                                                                                                                                                                             "type":  "object",
                                                                                                                                                                             "description":  "Event data"
                                                                                                                                                                         },
                                                                                                                                                                "source":  {
                                                                                                                                                                               "type":  "string",
                                                                                                                                                                               "description":  "Source service name",
                                                                                                                                                                               "example":  "test-service"
                                                                                                                                                                           }
                                                                                                                                                            }
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 },
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "Custom webhook sent"
                                                                                           }
                                                                               }
                                                             }
                                                }
              },
...
}